requirements.txt
requests==2.9.1
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
Pillow==2.3.0
django-ckeditor==5.0.3
decorator==3.4.0
pyimgur==0.5.3
dropbox==6.5.0
cloudinary==1.4.0
mezzanine
